server/requirements.txt
django
django_rest_framework
django-two-factor-auth
django-otp-yubikey